pyproject.toml
[project]
name = "summarization-poc"
version = "0.1.0"
description = ""
authors = [
    {name = "dailybotai",email = "171799305+dailybotai@users.noreply.github.com"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "pandas (>=2.3.1,<3.0.0)",
    "openai (>=1.95.0,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "streamlit (>=1.46.1,<2.0.0)",
    "altair (>=5.5.0,<6.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
